refactor(YALB-1687): split reponsibilities
