+++ web/profiles/custom/yalesites_profile/modules/custom/ys_templated_content/src/FileTypes/ContentTypedFilename.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\ys_templated_content\Support;
+namespace Drupal\ys_templated_content\FileTypes;
 
 /**
  * Helper to make templated filenames easier to retrieve data about.
