+++ web/profiles/custom/yalesites_profile/modules/custom/ys_templated_content/src/Helpers/TemplateFilenameHelper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\ys_templated_content\Support;
+namespace Drupal\ys_templated_content\Helpers;
 
 use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
